requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
executing==1.2.0
Flask==2.3.2
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.18.0
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
rsa==4.9
six==1.16.0
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
zipp==3.15.0
